+++ js/jquery.terminal-src.js
@@ -1069,9 +1069,19 @@
                 return value.done(callback);
             } else if (is_function(value.then)) {
                 return value.then(callback);
-            } else {
-                return callback(value);
+            } else if (value instanceof Array) {
+                var promises = value.filter(function(value) {
+                    return value && (is_function(value.done) || is_function(value.then));
+                });
+                if (promises.length) {
+                    var result = $.when.apply($, value).then(callback);
+                    if (is_function(value.catch)) {
+                        result.catch(error);
+                    }
+                    return result;
+                }
             }
+            return callback(value);
         }
     }
     // -----------------------------------------------------------------------
@@ -3931,7 +3941,7 @@
     function count_selfclosing_formatting(string) {
         var count = 0;
         if ($.terminal.have_formatting(string)) {
-            var re = new RegExp(format_parts_re, 'i');
+            var re = new RegExp(format_parts_re.source, 'i'); // without g flag
             $.terminal.format_split(string).forEach(function(str) {
                 if ($.terminal.is_formatting(str)) {
                     var m = str.match(re);
@@ -7164,8 +7174,15 @@
                     formatters: true,
                     convertLinks: settings.convertLinks
                 }, line.options || {});
-                var string;
-                string = stringify_value(line.value);
+                var string = stringify_value(line.value);
+                if (string && is_function(string.then)) {
+                    // handle function that return a promise #629
+                    return string.then(function(string) {
+                        process_line($.extend(line, {
+                            value: string
+                        }));
+                    });
+                }
                 if (string !== '') {
                     if (!line_settings.raw) {
                         if (line_settings.formatters) {
@@ -7249,6 +7266,8 @@
                     alert_exception('[Internal Exception(process_line)]', e);
                 }
             }
+            // is it work with unpromise that ignore undefined
+            return true;
         }
         // ---------------------------------------------------------------------
         // :: Update terminal lines
@@ -7299,14 +7318,15 @@
             }
             try {
                 output_buffer = [];
-                $.each(lines_to_show, function(i, line) {
-                    process_line(line);
+                unpromise(lines_to_show.map(function(line) {
+                    return process_line(line);
+                }), function() {
+                    if (!options.update) {
+                        command_line.before(detached_output); // reinsert output
+                    }
+                    self.flush(options);
+                    fire_event('onAfterRedraw');
                 });
-                if (!options.update) {
-                    command_line.before(detached_output); // reinsert output
-                }
-                self.flush(options);
-                fire_event('onAfterRedraw');
             } catch (e) {
                 if (is_function(settings.exceptionHandler)) {
                     settings.exceptionHandler.call(self, e, 'TERMINAL (redraw)');
@@ -8480,6 +8500,7 @@
             // :: Low Level method that overwrites interpreter
             // -------------------------------------------------------------
             set_interpreter: function(user_intrp, login) {
+                var defer = $.Deferred();
                 function overwrite_interpreter() {
                     self.pause(settings.softPause);
                     make_interpreter(user_intrp, login, function(result) {
@@ -8487,6 +8508,7 @@
                         var top = interpreters.top();
                         $.extend(top, result);
                         prepare_top_interpreter(true);
+                        defer.resolve();
                     });
                 }
                 if (is_function(login)) {
@@ -8500,7 +8522,7 @@
                 } else {
                     overwrite_interpreter();
                 }
-                return self;
+                return defer.promise();
             },
             // -------------------------------------------------------------
             // :: Show user greetings or terminal signature
@@ -9047,14 +9069,19 @@
                             if (options) {
                                 lines[line][1] = $.extend(lines[line][1], options);
                             }
-                            process_line({
+                            var next = process_line({
                                 value: value,
                                 index: line,
                                 options: lines[line][1]
                             });
-                            self.flush({
-                                scroll: false,
-                                update: true
+                            // process_line can return a promise
+                            // value is function that resolve to promise
+                            unpromise(next, function() {
+                                // trigger flush even if next is undefined
+                                self.flush({
+                                    scroll: false,
+                                    update: true
+                                });
                             });
                         });
                     }
@@ -9155,19 +9182,21 @@
                             if (render(value, locals)) {
                                 return self;
                             }
-                            process_line({
+                            var next = process_line({
                                 value: value,
                                 options: locals,
                                 index: lines.length
                             });
-                            // extended commands should be processed only
-                            // once in echo and not on redraw
-                            locals.exec = false;
-                            lines.push([value, locals]);
-                            if (locals.flush) {
-                                self.flush();
-                                fire_event('onAfterEcho', [arg]);
-                            }
+                            unpromise(next, function() {
+                                // extended commands should be processed only
+                                // once in echo and not on redraw
+                                locals.exec = false;
+                                lines.push([value, locals]);
+                                if (locals.flush) {
+                                    self.flush();
+                                    fire_event('onAfterEcho', [arg]);
+                                }
+                            });
                         });
                     } catch (e) {
                         // if echo throw exception we can't use error to
@@ -9875,6 +9904,9 @@
             if (is_function(value)) {
                 value = value();
             }
+            if (value && is_function(value.then)) {
+                return value.then(stringify_value);
+            }
             if (get_type(value) !== 'string') {
                 if (is_function(settings.parseObject)) {
                     var ret = settings.parseObject(value);
@@ -10450,12 +10482,12 @@
             function exec_spec(spec) {
                 var terminal = terminals.get()[spec[0]];
                 // execute if belong to this terminal
+                var defer = $.Deferred();
                 if (terminal && terminal_id === terminal.id()) {
                     if (!spec[2]) {
                         defer.resolve();
                         return defer.promise();
                     } else if (paused) {
-                        var defer = $.Deferred();
                         resume_callbacks.push(function() {
                             return terminal.exec(spec[2]).done(function() {
                                 terminal.save_state(spec[2], true, spec[1]);